close open sub-paths to make sure we warp implied self-closing line
There are a couple of flags (e.g. AS.svg) that contain a triangle composed of a move and two lines, without the closing 'Z'. For this reason the last implied line is not warped. Since we run topicosvg() all strokes have been converted to filled paths, so it is safe to assume all the paths are going to be filled only, and as such they will be rendered as if they had a Z at the end. We add that Z here, so in the prep_callback we can turn that into a line, which is then warped as usual.