translate and skewY gradients around mid point based on the flag warp
Fixes #6 sort of

We can't actually 'wave' the gradients' geometries (without mesh
gradients) so we cheat and simply apply a vertical translate and skew based on the direction of the warp around the mid-point of the gradient.
This seems to give satisfactory results for the few flags that do use gradients.