Add option to add semi-transparent border to flags
Partly fixes https://github.com/rsheeter/warp/issues/20

It turns out we don't actually need MULTIPLY compositing mode to stick a border around flags, normal alpha blending achieves the same visual effect (I suspect because the border is gray?). So we can add the border directly to the waved SVG flags (the soft-light we'll have to add it directly to the built COLRv1 table as post-processing step for SVG can't do soft-light -- or rather one could use CSS mix-blend-mode but it won't work for OT-SVG anyway so it's a dead end).